# src/weather_api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Weather API",
        "description": "API para consultar el clima y gestionar usuarios",
        "contact": {
            "email": "tuemail@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Listar todos los roles",
                "description": "Devuelve todos los roles registrados en el sistema",
                "operationId": "getRoles",
                "responses": {
                    "200": {
                        "description": "Lista de roles"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "Crear un nuevo rol",
                "description": "Crea un nuevo rol en el sistema",
                "operationId": "createRole",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "editor"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rol creado exitosamente"
                    },
                    "400": {
                        "description": "Datos inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/roles/{id}": {
            "delete": {
                "tags": [
                    "Roles"
                ],
                "summary": "Eliminar un rol",
                "description": "Elimina un rol por ID",
                "operationId": "deleteRole",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del rol",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rol eliminado correctamente"
                    },
                    "404": {
                        "description": "Rol no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/permissions": {
            "get": {
                "tags": [
                    "Permisos"
                ],
                "summary": "Listar todos los permisos",
                "description": "Devuelve todos los permisos registrados en el sistema",
                "operationId": "getPermissions",
                "responses": {
                    "200": {
                        "description": "Lista de permisos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Permisos"
                ],
                "summary": "Crear un nuevo permiso",
                "description": "Crea un nuevo permiso en el sistema",
                "operationId": "createPermission",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "edit post"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Permiso creado exitosamente"
                    },
                    "400": {
                        "description": "Datos inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/permissions/{id}": {
            "delete": {
                "tags": [
                    "Permisos"
                ],
                "summary": "Eliminar un permiso",
                "description": "Elimina un permiso por ID",
                "operationId": "deletePermission",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del permiso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permiso eliminado correctamente"
                    },
                    "404": {
                        "description": "Permiso no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/assign-role": {
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "Asignar un rol a un usuario",
                "description": "Asigna un rol específico a un usuario",
                "operationId": "assignRoleToUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "role"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rol asignado correctamente"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/assign-permission": {
            "post": {
                "tags": [
                    "Permisos"
                ],
                "summary": "Asignar un permiso a un rol",
                "description": "Asigna un permiso a un rol específico",
                "operationId": "assignPermissionToRole",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "role",
                                    "permission"
                                ],
                                "properties": {
                                    "role": {
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "permission": {
                                        "type": "string",
                                        "example": "view weather"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Permiso asignado al rol correctamente"
                    },
                    "404": {
                        "description": "Rol o permiso no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}